城市改成xlsx配置

+++ server/gameslg/data/init.go
@@ -4,7 +4,7 @@ import "github.com/llr104/LiFrame/server/gameslg/slgdb"
 
 func Init() {
 	CityMgr = cityManager{
-		cityMap:make(map[int] *slgdb.City),
+		cityMap:make(map[int16] *slgdb.City),
 	}
 	CityMgr.load()
 }
